requirements.txt
colorama>=0.4.6
requests>=2.31.0
dnspython>=2.4.0
python-whois>=0.9.4
flask>=3.0.0
